Allow rounds to be weighted for SOS calculations
In hypothetical tournaments which weight some rounds over others,
SOS calculations should take the round weighting into account.

For example, in a swiss tournament which includes some single-sided
rounds, those rounds would need to be weighted at 0.5 of a normal
round, due to the players in that round only playing a single game.