Abstract swiss pairing for large fields into a separate class
The BigSwiss pairing strategy will take a more performant but less
elegant approach to pairing a swiss field. Starting from the top, it
pairs each batch of players on equal points, randomly "overflowing"
one if the batch has an odd number. It can run into problems for:
small fields, fields that have lots of players on unique scores
(particularly at the bottom of the standings), and tournaments with
a large number of rounds. Even in these cases, the result will be a
player being paired against an opponent (or bye) again. This will
only affect players very near the bottom of standings, and only in
rounds.

In testing, the effect was minimal, even when it appeared, and will
not affect higher standings (i.e. cuts) except conceivably via SOS
or XSOS or the like, and not in a significant way (i.e. neither in
favour of or against any affected party).

The strategy is bad for small fields, but the "normal" graph method
works well for fields up to ~60 where this implementation takes over.