Improve page loads
Mostly improving the loading of the standings page which previously
would calculate standings on the fly on every page load. This was fine
for small events but is impractical and a huge problem for large events.

Now, the standings are calculated whenever a round is marked as
completed or a result is changed on a round that is already complete.
This in practice means the standings will be calculated just once per
round in most cases.

The standings page will now just display the cached standings (via the
StandingRow model) from the database and do no meaningful calculation.

Also restricted the pairings page to only show the last round for large
fields since that page can get very bloated and is high traffic.

The side effect of this is that you will not be able to see or amend
results from previous rounds in large events. This is probably fine but
could be something to remedy in the future.